port-serial: add HDLC command support
Add mm_port_serial_hdlc_command() which encapsulates the given data in
an HDLC frame and sends it to the port.  While the command is in the
queue, responses are expected as HDLC and unframed before sending to
subclass processing functions.

Also add property to the port itself that indicates responses are expected
to be HDLC framed, since an always-HDLC port like QCDM may receive
unsolicited responses while no command is being processed, but these
must be handled as HDLC.

The reason this isn't all encapsulated in MMPortSerialQcdm is that we
will want to use the AT port with HDLC framing for PPP in the future.