port-serial: split finalize into dispose and finalize
It seems signal handlers are disconnected by GLib during dispose,
so attempting to disconnect them in finalize() yields a GLib warning
about the signal ID not found on the instance.  So close the port
in dispose (which should be safe) and free other stuff in finalize.

(ModemManager:19003): GLib-GObject-WARNING **: gsignal.c:2641: instance '0x73e530' has no handler with id '191'