base-call: remove 'delete' operation
There is nothing to delete in the modem when a call is removed from
the call list, so there is no point in having such a method in the
call object. I assume this was a copy & paste leftover from the
MMSmsList/MMBaseSms implementation where we do have things to remove
(SMS parts stored in SMS storage) on such a delete operation.