shared-qmi: fix parent location interface management
We cannot rely on g_type_interface_peek_parent() during runtime to
give us the correct parent interface that we need to use, because if
either the MMBroadbandModemQmi or MMBroadbandModemMbim object gets
subclassed, this first parent interface reported would be different
than the one expected.

Fix this by explicitly storing the parent interface that we want
to use when the class is created.

This is the same logic already done for the Cinterion QMI modem object
that inherits from MMBroadbandModemQmi and implements custom location
management as well.